chore(): type fixes with verbatim syntax
