+++ src/Main.ts
@@ -1,6 +1,9 @@
+import Database from 'better-sqlite3';
+import { eq } from 'drizzle-orm';
 import { drizzle } from 'drizzle-orm/better-sqlite3';
 
 import { IAMClient } from './Client.js';
+import { user } from './Schema/Schema.js';
 
 export interface IAMClientOptions {
   databaseUrl: string;
